refactor and update README_Alice
