+++ app/controllers/sessions.rb
@@ -1,0 +1,24 @@
+class NomDiaries < Sinatra::Base
+
+  get '/sessions/new' do
+    erb :'/sessions/new'
+  end
+
+  post '/sessions' do
+    user = User.authenticate(params[:email], params[:password])
+    if user
+      session[:id] = user.id
+      redirect to('/noms')
+    else
+      flash.now[:errors] = ['The email or password is incorrect']
+      erb :'sessions/new'
+    end
+  end
+
+  delete '/sessions' do
+    session[:id] = nil
+    flash.keep[:notice] = 'goodbye!'
+    redirect to '/noms'
+  end
+
+end
