+++ app/views/_welcome.erb
@@ -1,0 +1,9 @@
+<%if current_user %>
+  Welcome, <%= current_user.nom_name%>
+
+  <form method='post' action='/sessions'>
+    <input type='hidden' name='_method' value='delete'>
+    <input type='submit' value='Sign out'>
+  </form>
+
+<%end%>
