+++ app/views/layout.erb
@@ -6,30 +6,8 @@
   </head>
 
   <body>
-    <%if current_user %>
-      Welcome, <%= current_user.nom_name%>
-
-      <form method='post' action='/sessions'>
-        <input type='hidden' name='_method' value='delete'>
-        <input type='submit' value='Sign out'>
-      </form>
-
-    <%end%>
-
-    <% if flash[:errors] && !flash[:errors].empty? %>
-      Please refer to the following errors below:
-      <ul id='errors'>
-        <% flash[:errors].each do |error| %>
-          <li><%= error %></li>
-        <% end %>
-    <% end %>
-
-    <% if flash[:notice] %>
-        <div id='notice'>
-          <%= flash[:notice] %>
-        </div>
-      <% end %>
-
+    <%= partial :welcome %>
+    <%= partial :flash %>
     <%= yield %>
   </body>
